feat(icons): adjust structure to devskills icons
