--- src/components/Icons/Business/style.js
@@ -1,6 +1,0 @@
-import styled from "styled-components"
-import Img from "gatsby-image"
-
-export const DevWrapper = styled(Img)`
-  width: 50px;
-`
